+++ src/scene/object/Eye.hpp
@@ -20,7 +20,7 @@ public:
 	Eye( Eye & other );
 	virtual ~Eye();
 
-	virtual void updateSelf( const float & delta );
+	virtual void updateSelf( const double & delta );
 	virtual void drawSelf();
 	virtual void drawSelfPost();
